Again, use intptr_t when casting from gpointer to int.

+++ contrib/bobtoolz/dialogs/dialogs-gtk.cpp
@@ -82,7 +82,7 @@ static void dialog_button_callback( GtkWidget *widget, gpointer data ){
 	ret = (int*)g_object_get_data( G_OBJECT( parent ), "ret" );
 
 	*loop = 0;
-	*ret = (int)data;
+	*ret = (intptr_t)data;
 }
 
 static gint dialog_delete_callback( GtkWidget *widget, GdkEvent* event, gpointer data ){
